+++ lib/win32_gui.dart
@@ -6,7 +6,7 @@ export 'dart:ffi';
 export 'package:win32/win32.dart';
 
 export 'src/components/win32_component_button.dart';
+export 'src/components/win32_richedit.dart';
 export 'src/win32_constants.dart';
 export 'src/win32_constants_extra.dart';
 export 'src/win32_gui_base.dart';
-export 'src/components/win32_richedit.dart';