+++ test/win32_gui_test.dart
@@ -138,24 +138,25 @@ class _MainWindow extends Window {
 
   @override
   void repaint(int hwnd, int hdc) {
-    super.repaint(hwnd, hdc);
-
     setIcon(iconDartLogoPath);
+    var hBitmap = loadImageCached(imageDartLogoPath);
+    var imgDimension = getBitmapDimension(hBitmap);
 
-    var imgW = 143;
-    var imgH = 139;
+    // Valid Bitmap:
+    if (imgDimension != null) {
+      var imgW = imgDimension.width;
+      var imgH = imgDimension.height;
 
-    var hBitmap = loadImageCached(imageDartLogoPath, imgW, imgH);
+      final hSpace = (dimensionWidth - imgW);
+      //final vSpace = (dimensionHeight - imgH);
+      final xCenter = hSpace ~/ 2;
+      //final yCenter = vSpace ~/ 2;
 
-    final hSpace = (dimensionWidth - imgW);
-    //final vSpace = (dimensionHeight - imgH);
-    final xCenter = hSpace ~/ 2;
-    //final yCenter = vSpace ~/ 2;
+      final x = xCenter;
+      final y = 10;
 
-    final x = xCenter;
-    final y = 10;
-
-    drawImage(hdc, hBitmap, x, y, imgW, imgH);
+      drawImage(hdc, hBitmap, x, y, imgW, imgH);
+    }
 
     textOutput.callRepaint();
   }
@@ -177,8 +178,6 @@ class _TextOutput extends RichEdit {
 
   @override
   void repaint(int hwnd, int hdc) {
-    super.repaint(hwnd, hdc);
-
     setBkColor(RGB(32, 32, 32));
     setTextColor(hdc, RGB(255, 255, 255));